Move calendar list entry into separate component
