+++ src/components/AppNavigation/Colorpicker.vue
